doublle -> float

+++ excercise2.cpp
@@ -5,7 +5,7 @@ using std::cout;
 
 struct Wheel
 {
-    double radius;
+    float radius;
     int angle;
 };
 
@@ -14,7 +14,7 @@ double getDistance(Wheel wheel);
 int main(void)
 {
     Wheel wheel = {4.5, 60};
-    double distanceTraveledByWheel;
+    float distanceTraveledByWheel;
 
     distanceTraveledByWheel = getDistance(wheel);
     cout << "distanceTraveledByWheel = " << distanceTraveledByWheel << "\n";
@@ -24,5 +24,5 @@ int main(void)
 
 double getDistance(Wheel wheel)
 {
-    return wheel.radius * 2 * M_PI * ((double)wheel.angle / 360);
+    return wheel.radius * 2 * M_PI * ((float)wheel.angle / 360);
 }
